Changed strategy of attribute values parsing.
In the past, a value which includes ':' has been parsed into XSD::QName.  And
when the value has been failed to be parsed as a qualified name (such as
'http://foo', or 'foo:bar' where the tag 'foo' is not defined), the value was
kept as a String.

Now, all attribute values are parsed with XSD::NS#parse, and XSD::NS#parse
returns XSD::QName.new(nil, '....') when the given String cannot be parsed as
a qualified name.  Then the value is kept as a QName.

Classes which accesses to attribute value follows this change.  Big change, but
I think it's the right way to go.