Trying to get the test suite running in ruby 2.2 Some techniques used in the test suite make them a little brittle.
Issue 1: Webrick + CGI scripts as test soap servers
Webrick starts the CGI script in a new environment,
without respecting rvm gemsets or bundler. Therefore
the necessary 'logger-application' gem in ruby 2.2
is not available. I tried to solve this, by looking
up the logger-application load path in the main 
process, and passing it as "-I ..." option to the
ruby bin used for the CGI script. Works now in 2.2,
does not break in 2.1. On the long run, CGI should
be removed from the tests, i think.

Issue 2: test/soap/marshal/mashaltestlib contains
marshalling test for the Time class which compare
marshalled/unmarshalled Time instances via musecs.
These fail once in a while (more often in ruby 2.2),
due to differences of 1musec. My guess is some kind
of rounding issue somewhere. I tried comparing
nsec, which is a little bit more reliable, but still
fragile. There must be a better way to compare the
Time objects.

Issue 3: In test/soap/test_envelopenamespace.rb the
Webrick test server takes a little to long to shutdown
and therefore blocks the 17171 port for the next test
(test_custommap). When that happens, all consequent
tests fail with "address already in use". I tried
debugging it, watching the Webrick status and the
server thread status. Webrick is getting stopped,
the server thread terminates, but it still blocks
the port for another few msecs. For now i have put
a `sleep(1)` behind the server termination, which
works fine, but is a little hacky. I can't figure out
why the Webrick shutdown is an issue only in this
particular test.